+++ test/unit/fuzz-tools.test.js
@@ -513,6 +513,7 @@ import * as capthusV535 from '../../lib/capthus-v535.js';
 import * as hardmanV536 from '../../lib/hardman-v536.js';
 import * as alsfrsRV537 from '../../lib/alsfrs-r-v537.js';
 import * as neosV538 from '../../lib/neos-v538.js';
+import * as islLymphedemaV539 from '../../lib/isl-lymphedema-v539.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -971,6 +972,7 @@ const MODULES = {
   'hardman-v536.js': hardmanV536,
   'alsfrs-r-v537.js': alsfrsRV537,
   'neos-v538.js': neosV538,
+  'isl-lymphedema-v539.js': islLymphedemaV539,
   'tb-testing.js': tbTesting,
 };